pyproject.toml
[tool.poetry]
name = "currency_conversion"
version = "0.1.0"
description = "currency_conversion"
authors = ["dias <@d.akhanbayev >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
alembic = "^1.13.1"
aiohttp = "^3.9.3"
pre-commit = "^3.7.0"
pydantic = "^2.6.4"
aiosignal = "^1.3.1"
aiosqlite = "^0.20.0"
annotated-types = "^0.6.0"
anyio = "^4.3.0"
astroid = "^3.1.0"
async-timeout = "^4.0.3"
asyncpg = "^0.29.0"
attrs = "^23.2.0"
certifi = "^2024.2.2"
cfgv = "^3.4.0"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
coverage = "^7.4.4"
dill = "^0.3.8"
distlib = "^0.3.8"
dnspython = "^2.6.1"
email-validator = "^2.1.1"
filelock = "^3.13.3"
flake8 = "^7.0.0"
frozenlist = "^1.4.1"
greenlet = "^3.0.3"
h11 = "^0.14.0"
httpcore = "^1.0.5"
httptools = "^0.6.1"
httpx = "^0.27.0"
identify = "^2.5.35"
idna = "^3.6"
iniconfig = "^2.0.0"
isort = "^5.13.2"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.3"
mako = "^1.3.2"
markupsafe = "^2.1.5"
mccabe = "^0.7.0"
multidict = "^6.0.5"
nodeenv = "^1.8.0"
orjson = "^3.10.0"
packaging = "^24.0"
platformdirs = "^4.2.0"
pluggy = "^1.4.0"
pycodestyle = "^2.11.1"
pydantic-extra-types = "^2.6.0"
pydantic-settings = "^2.2.1"
pydantic-core = "^2.18.0"
pyflakes = "^3.2.0"
pylint = "^3.1.0"
pytest = "8.0.2"
pytest-asyncio = "0.23.5"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
pyyaml = "^6.0.1"
requests = "^2.31.0"
sniffio = "^1.3.1"
sqlalchemy = "^2.0.29"
starlette = "^0.37.2"
tomlkit = "^0.12.4"
typing-extensions = "^4.10.0"
ujson = "^5.9.0"
urllib3 = "^2.2.1"
uvicorn = "^0.29.0"
uvloop = "^0.19.0"
virtualenv = "^20.25.1"
watchfiles = "^0.21.0"
websockets = "^12.0"
yarl = "^1.9.4"
bandit = "^1.7.8"
celery = "^5.4.0"
redis = "^5.0.5"
onnxruntime = "1.15.1"
pillow = "^10.3.0"
ffmpeg-python = "^0.2.0"
numpy = "1.26.4"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
